kb/batch_embedding/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0
cohere==5.16.1
pinecone==7.3.0
pydantic>=2.0.0
PyPDF2>=3.0.0
pypdf>=3.17.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
numpy>=1.24.0
httpx>=0.24.0
fastapi>=0.104.0
pdfplumber>=0.11.0
